+++ CHANGELOG.md
@@ -49,12 +49,12 @@ ALTER TABLE authclients ADD COLUMN "clientDisabled" boolean NOT NULL DEFAULT FAL
 server/validate.js - The `validate.client()` function was updated to fail validation 
 when clientDisabled property is set to true. This will deny access to API routes
 that use Basic Auth with base64 encoded client credentials. These are:
-POST /oauth/token, POST /oauth/token/introspect, POST /oauth/token/revoke.
+POST /oauth/token, POST /oauth/introspect, POST /oauth/token/revoke.
 
 server/validate.js - The `validate.token()` function was updated to fail validation 
 when clientDisabled property is set to true for the client account that created
 the access token. This will return an error for API route 
-POST /oauth/token/introspect, used to validate submitted tokens.
+POST /oauth/introspect, used to validate submitted tokens.
 
 server/oauth2.js - Those oauth2orize callback functions have been updated 
 to return a status 401 error to outh2orize when the client.clientDisabled